+++ js/data.js
@@ -1,0 +1,19 @@
+const PHOTO_DESCRIPTIONS = [
+  'Случайное описание 1',
+  'Случайное описание 2',
+  'Случайное описание 3'
+];
+
+const COMMENT_MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.'
+];
+
+const COMMENT_NAMES = [
+  'Kate',
+  'Nick',
+  'Alex',
+  'John'
+];
+
+export {PHOTO_DESCRIPTIONS, COMMENT_MESSAGES, COMMENT_NAMES};
